feat(Screens): remove container & add custom hooks
